+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/data/work/CacheBookWork.kt
@@ -1,6 +1,7 @@
 package indi.dmzz_yyhyy.lightnovelreader.data.work
 
 import android.content.Context
+import android.os.Looper
 import android.widget.Toast
 import androidx.hilt.work.HiltWorker
 import androidx.work.Worker
@@ -31,27 +32,29 @@ class CacheBookWork @AssistedInject constructor(
         if (bookId < 0) return Result.failure()
         val downloadItem = MutableDownloadItem(DownloadType.CACHE, bookId)
         downloadProgressRepository.addExportItem(downloadItem)
-        var total: Int
         var count = 0
-        webBookDataSource.getBookVolumes(bookId)?.let { bookVolumes ->
-            total = bookVolumes.volumes.sumOf { it.chapters.size } + 1
-            if (bookVolumes.volumes.isEmpty()) return Result.failure()
-            localBookDataSource.updateBookVolumes(bookId, bookVolumes)
-            bookVolumes.volumes.forEach { volume ->
-                volume.chapters.map { it.id }.forEach { chapterId ->
-                    val chapter = webBookDataSource.getChapterContent(
-                        chapterId = chapterId,
-                        bookId = bookId
-                    ) ?: return Result.failure()
-                    if (chapter.isEmpty()) return Result.failure().also { Toast.makeText(applicationContext, "缓存失败，请检查网络环境", Toast.LENGTH_SHORT).show() }
-                    localBookDataSource.updateChapterContent(chapter)
-                    count ++
-                    downloadItem.progress = count.toFloat()/total
+        val bookVolumes = webBookDataSource.getBookVolumes(bookId)
+        val total = bookVolumes.volumes.sumOf { it.chapters.size } + 1
+        if (bookVolumes.volumes.isEmpty()) return Result.failure()
+        localBookDataSource.updateBookVolumes(bookId, bookVolumes)
+        bookVolumes.volumes.forEach { volume ->
+            volume.chapters.map { it.id }.forEach { chapterId ->
+                val chapter = webBookDataSource.getChapterContent(
+                    chapterId = chapterId,
+                    bookId = bookId
+                )
+                if (chapter.isEmpty()) return Result.failure().also {
+                    Looper.prepare()
+                    Toast.makeText(applicationContext, "缓存失败，请检查网络环境", Toast.LENGTH_SHORT).show()
+                    Looper.loop()
                 }
+                localBookDataSource.updateChapterContent(chapter)
+                count ++
+                downloadItem.progress = count.toFloat() / total
             }
         }
         webBookDataSource.getBookInformation(bookId)
-            ?.let {
+            .let {
                 if (it.isEmpty()) return Result.failure()
                 localBookDataSource.updateBookInformation(it)
             }
