+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/book/detail/ExportSettings.kt
@@ -1,0 +1,12 @@
+package indi.dmzz_yyhyy.lightnovelreader.ui.book.detail
+
+data class ExportSettings(
+    val selectedVolumeIds: Set<Int> = emptySet(),
+    val includeImages: Boolean = true,
+    val exportType: ExportType = ExportType.BOOK
+)
+
+enum class ExportType {
+    BOOK,
+    VOLUMES
+}
