+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/book/reader/content/scroll/ScrollContentViewModel.kt
@@ -34,7 +34,6 @@ class ScrollContentViewModel(
     )
 
     init {
-        println("ViewModelInit")
         coroutineScope.launch(Dispatchers.IO) {
             settingState.isUsingContinuousScrollingUserData.getFlowWithDefault(false).collect {
                 if (it) {
